Magnitude of completeness

Note

maxcurv, mbass and mbs are based on magnitudes grouped into bins of width $ΔM$. As a result, the estimated magnitude of completeness $M_{c}$ corresponds to the center of one of the bins. To obtain the effective magnitude of completeness $M_{0}$, it is then necessary to substract from $M_{c}$ half bin width: $M_{0} = M_{c} - ΔM/2$

MagnitudeDistributions.maxcurv — Function

maxcurv(mags, ΔM)

Apply the maximum curvature method to compute the magnitude of completeness.

Arguments

mags: array storing the magnitude values of an incomplete catalog.

ΔM: bin width to be used to group the magnitude values.`

maxcurv(bincenters, binevents)

Apply the maximum curvature method to compute the magnitude of completeness.

Arguments

bincenters: array storing the central values of the magnitude bins.

binevents: array storing the number of events in the magnitude bins.

bincenters and binevents can be obtained with histmag(mag, ΔM).

MagnitudeDistributions.mbass — Function

mbass(mags, ΔM)

Apply the Median-Based Analysis of the Segment Slope method to estimate the magnitude of completeness.

Arguments

mags: array storing the magnitude values of an incomplete catalog.

ΔM: bin width to be used to group the magnitude values.`

References

Amorese 2007. Applying a Change-Point Detection Method on Frequency-Magnitude Distributions. Bulletin of the Seismological Society of America 97 (5): 1742–1749. https://doi.org/10.1785/0120060181

Lanzante 1996. Resistant, Robust and Non‐Parametric Techniques for the Analysis of Climate Data: Theory and Examples, Including Applications to Historical Radiosonde Station Data. Int. J. Climatol., 16: 1197-1226. https://doi.org/10.1002/(SICI)1097-0088(199611)16:11<1197::AID-JOC89>3.0.CO;2-L

mbass(bincenters, binevents)

Apply the Median-Based Analysis of the Segment Slope method to estimate the magnitude of completeness.

Arguments

bincenters: array storing the central values of the magnitude bins.

binevents: array storing the number of events in the magnitude bins.

bincenters and binevents can be obtained with histmag(mag, ΔM).

MagnitudeDistributions.mbs — Function

mbs(mags, ΔM; winsize=5, minevents=50)

Compute the magnitude of completeness with the b−value stability (MBS) method.

Arguments

mags: array storing the magnitude values of an incomplete catalog.

ΔM: bin width to be used to group the magnitude values.`

Keyword arguments

winsize: window length used for the moving average. Default is 5 magnitude bins

minevents: minimum number of events allowed when calculating b-values. Defaul is 50 events.

References:

Cao and Gao, 2002, Temporal variations of seismic b-values beneath northeastern japan island arc, Geophysical Research Letters, 29, https://doi.org/10.1029/2001GL013775

Woessner and Wiemer (2005), Assessing the quality of earthquake catalogues: Estimating the magnitude of completeness and its uncertainty, Bulletin of the Seismological Society of America, 95, https://doi.org/10.1785/0120040007